Refactor app startup and global resources
