+++ LoloRecorder/App.xaml.cs
@@ -1,8 +1,15 @@
 using System.Windows;
+using LoloRecorder.Views;
 
 namespace LoloRecorder
 {
     public partial class App : Application
     {
+        protected override void OnStartup(StartupEventArgs e)
+        {
+            base.OnStartup(e);
+            var mainWindow = new MainWindow();
+            mainWindow.Show();
+        }
     }
 }
