+++ LoloRecorder/Resources/Colors.xaml
@@ -1,0 +1,5 @@
+<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    <SolidColorBrush x:Key="Fundo" Color="#1e1e1e"/>
+    <SolidColorBrush x:Key="Acento" Color="#0a84ff"/>
+</ResourceDictionary>
