+++ LoloRecorder/Views/MainWindow.xaml
@@ -1,13 +1,9 @@
-<Window x:Class="LoloRecorder.MainWindow"
+<Window x:Class="LoloRecorder.Views.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="LoloRecorder" Height="350" Width="525">
     <Window.Resources>
         <ResourceDictionary>
-            <!-- Global color resources -->
-            <SolidColorBrush x:Key="Fundo" Color="#1e1e1e"/>
-            <SolidColorBrush x:Key="Acento" Color="#0a84ff"/>
-
             <!-- Global button style -->
             <Style TargetType="Button">
                 <Setter Property="Background" Value="{StaticResource Acento}"/>
